NO.1 Backend core specifications and contract establishment
A Unified API response format is established, DTO (data transfer object) is introduced to cut off the direct coupling between Controller and database entity, and a global exception handling mechanism is established.

+++ finance-tracker/src/main/java/com/gcc_0119/finance_tracker/model/Transaction.java
@@ -4,6 +4,7 @@ import lombok.Data;
 import org.springframework.data.annotation.Id;
 import org.springframework.data.mongodb.core.mapping.Document;
 
+import java.math.BigDecimal; 
 import java.time.LocalDateTime;
 
 @Document(collection = "transactions")
@@ -14,7 +15,8 @@ public class Transaction {
     private String type;        
     private String category;    
 
-    private Double amount;
+    private BigDecimal amount;
+    
     private String description; 
     private LocalDateTime timestamp;
